+++ usr/src/uts/common/io/nxge/nxge_txdma.c
@@ -19,7 +19,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2008 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2009 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -1075,17 +1075,14 @@ nxge_txdma_reclaim(p_nxge_t nxgep, p_tx_ring_t tx_ring_p, int nmblks)
 			tx_msg_p = &tx_msg_ring[tx_rd_index];
 		}
 
-		status = (nmblks <= (tx_ring_p->tx_ring_size -
-		    tx_ring_p->descs_pending -
-		    TX_FULL_MARK));
+		status = (nmblks <= ((int)tx_ring_p->tx_ring_size -
+		    (int)tx_ring_p->descs_pending - TX_FULL_MARK));
 		if (status) {
 			cas32((uint32_t *)&tx_ring_p->queueing, 1, 0);
 		}
 	} else {
-		status = (nmblks <=
-		    (tx_ring_p->tx_ring_size -
-		    tx_ring_p->descs_pending -
-		    TX_FULL_MARK));
+		status = (nmblks <= ((int)tx_ring_p->tx_ring_size -
+		    (int)tx_ring_p->descs_pending - TX_FULL_MARK));
 	}
 
 	NXGE_DEBUG_MSG((nxgep, TX_CTL,
